pyproject.toml
[project]
name = "chatrobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "fastapi>=0.121.3",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "pydantic-settings>=2.0.0",
    "pymysql>=1.1.2",
    "aiomysql>=0.2.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
    "python-docx>=1.2.0",
    "pypdf>=6.4.0",
    "python-pptx>=1.0.2",
    "langchain-text-splitters>=1.0.0",
]
